pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "beancount-lalitm"
version = "0.1.0"
description = "Importers and plugins for Beancount plain-text accounting"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    { name = "Lalit Maganti" }
]
keywords = ["beancount", "accounting", "finance", "importers", "plugins"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Accounting",
]
dependencies = [
    "beancount>=3.0.0",
    "beangulp>=0.1.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Importers with additional dependencies
ib = ["xmltodict>=0.13.0"]
pdf = ["pdftotext>=2.2.0"]
excel = ["openpyxl>=3.1.0"]
all = [
    "xmltodict>=0.13.0",
    "pdftotext>=2.2.0",
    "openpyxl>=3.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/LalitMaganti/beancount-lalitm"
Repository = "https://github.com/LalitMaganti/beancount-lalitm"

[tool.hatch.build.targets.wheel]
packages = ["src/beancount_tools"]
